[Consensus Observer] Separate consensus observer state.
