+++ consensus/src/consensus_observer/observer/mod.rs
@@ -1,6 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod active_state;
 pub mod consensus_observer;
 pub mod ordered_blocks;
 pub mod payload_store;
